refactor(lib): Barrel files renaming
