+++ projects/igniteui-angular/src/lib/banner/banner.component.ts
@@ -2,7 +2,7 @@ import { Component, NgModule, EventEmitter, Output, Input, ViewChild, ElementRef
     ContentChild, HostBinding } from '@angular/core';
 import { IgxExpansionPanelModule } from '../expansion-panel/expansion-panel.module';
 import { AnimationSettings } from '../expansion-panel/expansion-panel.component';
-import { IgxExpansionPanelComponent } from '../expansion-panel';
+import { IgxExpansionPanelComponent } from '../expansion-panel/public_api';
 import { IgxIconModule, IgxIconComponent } from '../icon/index';
 import { IToggleView } from '../core/navigation';
 import { IgxButtonModule } from '../directives/button/button.directive';
