+++ projects/igniteui-angular/src/lib/calendar/calendar-multi-view.component.spec.ts
@@ -4,7 +4,7 @@ import { FormsModule } from '@angular/forms';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { configureTestSuite } from '../test-utils/configure-suite';
 import { UIInteractions } from '../test-utils/ui-interactions.spec';
-import { IgxCalendarComponent, IgxCalendarModule } from './index';
+import { IgxCalendarComponent, IgxCalendarModule } from './public_api';
 import { IgxDatePickerComponent, IgxDatePickerModule } from '../date-picker/date-picker.component';
 import { DateRangeType } from '../core/dates';
 import { HelperTestFunctions } from './calendar-helper-utils';
