+++ projects/igniteui-angular/src/lib/chips/public_api.ts
