+++ projects/igniteui-angular/src/lib/expansion-panel/public_api.ts
