+++ projects/igniteui-angular/src/lib/input-group/input-group.component.spec.ts
@@ -3,11 +3,11 @@ import { async, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { IgxInputGroupComponent, IgxInputGroupModule } from './input-group.component';
 import { DisplayDensityToken, DisplayDensity } from '../core/displayDensity';
-import { wait, UIInteractions } from '../test-utils/ui-interactions.spec';
+import { UIInteractions } from '../test-utils/ui-interactions.spec';
 import { IgxIconModule } from '../icon';
 import { IgxInputDirective } from '../directives/input/input.directive';
 import { configureTestSuite } from '../test-utils/configure-suite';
-import { IgxPrefixDirective, IgxSuffixDirective } from '../chips';
+import { IgxPrefixDirective, IgxSuffixDirective } from '../chips/public_api';
 
 const INPUT_GROUP_CSS_CLASS = 'igx-input-group';
 const INPUT_GROUP_BOX_CSS_CLASS = 'igx-input-group--box';
