+++ projects/igniteui-angular/src/lib/test-utils/grid-functions.spec.ts
@@ -4,7 +4,7 @@ import { By } from '@angular/platform-browser';
 import { ComponentFixture, tick } from '@angular/core/testing';
 import { IgxInputDirective } from '../input-group';
 import { IgxGridHeaderComponent } from '../grids/headers/grid-header.component';
-import { IgxChipComponent } from '../chips';
+import { IgxChipComponent } from '../chips/public_api';
 import { IgxGridComponent } from '../grids/grid/grid.component';
 import { IgxColumnGroupComponent } from '../grids/columns/column-group.component';
 import { IgxGridHeaderGroupComponent } from '../grids/headers/grid-header-group.component';
